update folder and file service

+++ src/files/files.service.ts
@@ -4,6 +4,7 @@ import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
 import { CreateFileInput } from './inputs/create.input';
 import { UpdateFileInput } from './inputs/update.input';
+import { GetManyFilesInput } from './inputs/get-many.input';
 
 @Injectable()
 export class FilesService {
@@ -18,13 +19,16 @@ export class FilesService {
     return res as unknown as File;
   }
 
-  async getOne(_id: string): Promise<File> {
-    return this.fileModel.findOne({ _id: _id }).exec() as unknown as File;
+  async getOne(_id: string, user_id: string): Promise<File> {
+    return this.fileModel
+      .findOne({ _id: _id, user_id: user_id })
+      .exec() as unknown as File;
   }
 
-  async getMany(page = 1, limit = 10): Promise<File[]> {
+  async getMany(params: GetManyFilesInput): Promise<File[]> {
+    const { limit, page, user_id } = params;
     const res = this.fileModel
-      .find()
+      .find({ user_id: user_id })
       .limit(limit)
       .skip((page - 1) * limit)
       .exec();
@@ -33,15 +37,16 @@ export class FilesService {
 
   async updateOne(
     _id: string,
-    data: Omit<UpdateFileInput, '_id'>,
+    user_id: string,
+    data: Omit<UpdateFileInput, '_id' | 'user_id'>,
   ): Promise<File> {
     return this.fileModel.findByIdAndUpdate(
-      { id: _id },
+      { id: _id, user_id: user_id },
       { ...data },
     ) as unknown as File;
   }
 
-  async deleteOne(_id: string): Promise<void> {
-    this.fileModel.findByIdAndDelete({ id: _id });
+  async deleteOne(_id: string, user_id: string): Promise<void> {
+    this.fileModel.findByIdAndDelete({ id: _id, user_id });
   }
 }
