+++ src/folders/folders.service.ts
@@ -4,6 +4,8 @@ import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
 import { CreateFolderInput } from './inputs/create.input';
 import { UpdateFolderInput } from './inputs/update.input';
+import { GetManyFolderInput } from './inputs/get-many.input';
+import { DeleteOneFolderInput } from './inputs/delete-one.input';
 
 @Injectable()
 export class FoldersService {
@@ -18,14 +20,16 @@ export class FoldersService {
     return res as unknown as Folder;
   }
 
-  async getOne(_id: string): Promise<Folder> {
-    return this.folderModel.findOne({ _id: _id }).exec() as unknown as Folder;
+  async getOne(_id: string, user_id: string): Promise<Folder> {
+    return this.folderModel
+      .findOne({ _id: _id, user_id: user_id })
+      .exec() as unknown as Folder;
   }
 
-  async getMany(page = 1, limit = 10): Promise<Folder[]> {
-    // get many
+  async getMany(params: GetManyFolderInput): Promise<Folder[]> {
+    const { page, limit, user_id } = params;
     const res = this.folderModel
-      .find()
+      .find({ user_id: user_id })
       .limit(limit)
       .skip((page - 1) * limit)
       .exec();
@@ -35,15 +39,17 @@ export class FoldersService {
 
   async updateOne(
     _id: string,
-    data: Omit<UpdateFolderInput, '_id'>,
+    user_id: string,
+    data: Omit<UpdateFolderInput, '_id' | 'user_id'>,
   ): Promise<Folder> {
     return this.folderModel.findByIdAndUpdate(
-      { id: _id },
+      { id: _id, user_id: user_id },
       { ...data },
     ) as unknown as Folder;
   }
 
-  async deleteOne(_id: string): Promise<void> {
-    this.folderModel.findByIdAndDelete({ id: _id });
+  async deleteOne(params: DeleteOneFolderInput): Promise<void> {
+    const { _id, user_id } = params;
+    this.folderModel.findByIdAndDelete({ id: _id, user_id: user_id });
   }
 }
